Add module ColorTransfer.CLI .
Allow choosing between ellipsoid transformation and histogram matching from CLI.
